# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nexum-backend",
    "description": "Sistema operativo financiero personal impulsado por IA.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/intelligence/snapshot": {
      "get": {
        "tags": [
          "intelligence"
        ],
        "summary": "Get Snapshot",
        "operationId": "get_snapshot_api_v1_intelligence_snapshot_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceSnapshotRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/intelligence/balance": {
      "get": {
        "tags": [
          "intelligence"
        ],
        "summary": "Get Balance",
        "operationId": "get_balance_api_v1_intelligence_balance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligenceBalanceRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountBalanceRead": {
        "properties": {
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "account_name": {
            "type": "string",
            "title": "Account Name"
          },
          "account_type": {
            "type": "string",
            "title": "Account Type"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "balance": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Balance"
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "account_name",
          "account_type",
          "currency",
          "balance"
        ],
        "title": "AccountBalanceRead"
      },
...
      "IntelligenceBalanceRead": {
        "properties": {
          "total_available_real": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Available Real"
          },
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountBalanceRead"
            },
            "type": "array",
            "title": "Accounts"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          }
        },
        "type": "object",
        "required": [
          "total_available_real",
          "accounts",
          "currency"
        ],
        "title": "IntelligenceBalanceRead"
      },
...
      "IntelligenceSnapshotRead": {
        "properties": {
          "period": {
            "$ref": "#/components/schemas/SnapshotPeriod"
          },
          "cash": {
            "$ref": "#/components/schemas/SnapshotCash"
          },
          "cashflow": {
            "$ref": "#/components/schemas/SnapshotCashflow"
          },
          "debt": {
            "$ref": "#/components/schemas/SnapshotDebt"
          },
          "goals": {
            "$ref": "#/components/schemas/SnapshotGoals"
          },
          "obligations": {
            "$ref": "#/components/schemas/SnapshotObligations"
          },
          "transfers": {
            "$ref": "#/components/schemas/SnapshotTransfers"
          },
          "recent_activity": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Recent Activity"
          }
        },
        "type": "object",
        "required": [
          "period",
          "cash",
          "cashflow",
          "debt",
          "goals",
          "obligations",
          "transfers",
          "recent_activity"
        ],
        "title": "IntelligenceSnapshotRead"
      },
...
      "SnapshotCash": {
        "properties": {
          "total_balance": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Balance"
          },
          "active_accounts_count": {
            "type": "integer",
            "title": "Active Accounts Count"
          }
        },
        "type": "object",
        "required": [
          "total_balance",
          "active_accounts_count"
        ],
        "title": "SnapshotCash"
      },
      "SnapshotCashflow": {
        "properties": {
          "income": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Income"
          },
          "expenses": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Expenses"
          },
          "net_cashflow": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Net Cashflow"
          }
        },
        "type": "object",
        "required": [
          "income",
          "expenses",
          "net_cashflow"
        ],
        "title": "SnapshotCashflow"
      },
      "SnapshotDebt": {
        "properties": {
          "credit_card_total_debt": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Credit Card Total Debt"
          },
          "billed_debt": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Billed Debt"
          },
          "unbilled_debt": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Unbilled Debt"
          }
        },
        "type": "object",
        "required": [
          "credit_card_total_debt",
          "billed_debt",
          "unbilled_debt"
        ],
        "title": "SnapshotDebt"
      },
      "SnapshotGoals": {
        "properties": {
          "active_goals_count": {
            "type": "integer",
            "title": "Active Goals Count"
          },
          "total_target": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Target"
          },
          "total_saved": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Saved"
          }
        },
        "type": "object",
        "required": [
          "active_goals_count",
          "total_target",
          "total_saved"
        ],
        "title": "SnapshotGoals"
      },
      "SnapshotObligations": {
        "properties": {
          "pending_count": {
            "type": "integer",
            "title": "Pending Count"
          },
          "pending_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Pending Amount"
          }
        },
        "type": "object",
        "required": [
          "pending_count",
          "pending_amount"
        ],
        "title": "SnapshotObligations"
      },
      "SnapshotPeriod": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          }
        },
        "type": "object",
        "required": [
          "month",
          "timezone"
        ],
        "title": "SnapshotPeriod"
      },
      "SnapshotTransfers": {
        "properties": {
          "monthly_transfer_volume": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Monthly Transfer Volume"
          }
        },
        "type": "object",
        "required": [
          "monthly_transfer_volume"
        ],
        "title": "SnapshotTransfers"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}